Rename: 상품 상세 페이지 파일명/컴포넌트명 변경(detail->products)
