+++ README.md
@@ -74,7 +74,7 @@ feature/#이슈번호
  ┃ ┣ 📂layout ➡️ 공통 컴포넌트
  ┃ ┗ 📂template ➡️ 특정 페이지에서만 쓰이는 컴포넌트
  ┃ ┃ ┣ 📂cart
- ┃ ┃ ┣ 📂detail
+ ┃ ┃ ┣ 📂products
  ┃ ┃ ┣ 📂join
  ┃ ┃ ┣ 📂mypage
  ┃ ┃ ┗ 📂payment
@@ -82,7 +82,7 @@ feature/#이슈번호
  ┣ 📂interfaces ➡️ 타입지정 관련
  ┣ 📂pages ➡️ 탬플릿에 작성된 컴포넌트를 import해서 구성
  ┃ ┣ 📂cart
- ┃ ┣ 📂detail
+ ┃ ┣ 📂products
  ┃ ┣ 📂main
  ┃ ┣ 📂mypage
  ┃ ┗ 📂payment
