+++ src/components/template/products/AccommodationInfo.tsx
@@ -3,7 +3,6 @@ import { handleCopyClipBoard } from '@/util/clipboard';
 import { useState } from 'react';
 import { GuestCount } from '@/interfaces/interface';
 import { GoHeart, GoShareAndroid } from 'react-icons/go';
-import APIServiceList from './APIServiceList';
 
 import {
   StyledIconBox,
@@ -12,7 +11,7 @@ import {
   StyledServiceInfo,
   StyledTextBox,
   StyledWrap,
-} from '@/style/detail/detailStyle';
+} from '@/style/products/productsStyle';
 import {
   StyledTitle,
   StyledText,
@@ -21,6 +20,7 @@ import {
 } from '@/style/payment/paymentStyle';
 import { Moment } from 'moment';
 import CalenderModal from '@/components/layout/modal/calenderModal';
+import ProductsFacilityList from './ProductsFacilityList';
 
 interface AccommodationProp {
   onOpen: (e: React.MouseEvent) => void;
@@ -75,7 +75,7 @@ const AccommodationInfo = ({
           $flexDirection="row"
           $justifyContent="flex-start"
           $gap="1rem">
-          <APIServiceList />
+          <ProductsFacilityList />
         </StyledServiceInfo>
         <StyledOnClick $color="#444" $borderBottom="none">
           ★4.50 후기 0개
