+++ src/components/template/products/ImageSlideModal.tsx
