--- src/pages/detail/detail.tsx
@@ -1,14 +1,0 @@
-import DetailContainer from '@/components/template/detail/DetailContainer';
-import { useParams } from 'react-router-dom';
-
-const Detail = () => {
-  const { accomodationID } = useParams<{ accomodationID: string }>();
-
-  return (
-    <div>
-      {accomodationID && <DetailContainer accomodationID={accomodationID} />}
-    </div>
-  );
-};
-
-export default Detail;
